env_site/summerjobportal/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.20
botocore==1.31.20
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.4
django-allauth==0.54.0
django-autoslug==1.9.9
django-cleanup==8.0.0
django-countries-with-calling-codes==0.1.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-elasticsearch-dsl==7.3
django-filter==23.2
django-pwa==1.1.0
django-smart-selects==1.6.0
django-storages==1.13.2
django-utils-six==2.0
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
filelock==3.12.2
frozenlist==1.4.0
geoip2==4.7.0
graphene==3.3
graphene-django==3.1.3
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
jmespath==1.0.1
maxminddb==2.4.0
multidict==6.0.4
oauthlib==3.2.2
platformdirs==3.10.0
promise==2.3
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==1.26.16
virtualenv==20.24.2
yarl==1.9.2
